backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-socketio==5.9.0
websockets==11.0.3
#aiohttp==3.8.5
